requirements.txt
waitress
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain_community
Flask
requests
jsonify